refactor: rename package domain.dto -> domain.vo
