+++ src/main/java/kitchenpos/eatinorders/tobe/domain/domainservice/MenuDomainService.java
@@ -1,7 +1,7 @@
 package kitchenpos.eatinorders.tobe.domain.domainservice;
 
 import java.util.Map;
-import kitchenpos.eatinorders.tobe.domain.dto.MenuRes;
+import kitchenpos.eatinorders.tobe.domain.vo.MenuRes;
 
 public interface MenuDomainService {
