+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrderTest.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import kitchenpos.eatinorders.tobe.domain.dto.MenuRes;
+import kitchenpos.eatinorders.tobe.domain.vo.MenuRes;
 import kitchenpos.eatinorders.tobe.domain.domainservice.OrderTableDomainService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
